Removed old JSON Importer class and renamed new class for consistency.
